requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==2.6.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
asn1crypto==1.4.0
beautifulsoup4==4.9.1
billiard==3.6.4.0
blinker==1.4
bson==0.5.10
celery==4.4.7
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
click==7.1.2
cryptography==2.7
flask-cors==3.0.9
flask==1.1.2
gevent==21.1.2
greenlet==1.1.1; platform_python_implementation == 'CPython'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
kombu==4.6.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==2.0.1; python_version >= '3.6'
marshmallow==3.2.1
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==1.7.1
pympler==0.9
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-slugify==4.0.1
pytz==2021.1
pyyaml==5.3.1
redis==3.5.3
requests==2.24.0
sentry-sdk==1.3.0
shapely==1.7.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.2.1; python_version >= '3.6'
text-unidecode==1.3
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uwsgi==2.0.19.1
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
werkzeug==2.0.1; python_version >= '3.6'
zope.event==4.5.0
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
